requirements.txt
Flask==0.12.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.6
Flask-Cors==3.0.3
gunicorn=19.7.1
Keras==2.1.5
numpy==1.14.2
tensorflow-gpu==1.4.0
